pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartcat"
version = "0.1.0"
description = "RAG-powered email search with AI agent"
requires-python = ">=3.11"
dependencies = [
    "python-dateutil>=2.8",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8",
    "sentence-transformers>=3.0",
    "qdrant-client>=1.9",
    "click>=8.1",
    "rich>=13.0",
    "structlog>=24.0",
    "tqdm>=4.66",
    "requests>=2.31",
    "einops>=0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
llm = [
    "llama-cpp-python>=0.2",
]
docling = [
    "docling>=2.0",
]
api = [
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "httpx>=0.27",
]
ocr = [
    "surya-ocr",
    "easyocr",
    "pymupdf",
]
translation = [
    "argostranslate",
    "langdetect",
]

[project.scripts]
smartcat = "smartcat.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (downloads models, deselect with '-m \"not slow\"')",
]
